+++ src/main/java/com/xiliulou/electricity/vo/merchant/RebateConfigVO.java
@@ -2,6 +2,8 @@ package com.xiliulou.electricity.vo.merchant;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author zzlong
  * @email zhaozhilong@xiliulou.com
@@ -31,22 +33,22 @@ public class RebateConfigVO {
     /**
      * 渠道员拉新返现
      */
-    private Double channelerInvitation;
+    private BigDecimal channelerInvitation;
     
     /**
      * 渠道员续费返现
      */
-    private Double channelerRenewal;
+    private BigDecimal channelerRenewal;
     
     /**
      * 商户拉新返现
      */
-    private Double merchantInvitation;
+    private BigDecimal merchantInvitation;
     
     /**
      * 商户续费返现
      */
-    private Double merchantRenewal;
+    private BigDecimal merchantRenewal;
     
     /**
      * 状态 0:关闭,1:开启
